style(eslint): disable type assertions and class member spacing rules

+++ src/domain/application/usecases/delete-account/delete-account.usecase.ts
@@ -2,6 +2,7 @@ import { Inject, Injectable } from '@nestjs/common'
 import { UseCase } from '@/core/domain/application/use-case'
 import { RefreshTokensRepository } from '@/domain/application/repositories/refresh-tokens.repository'
 import { UsersRepository } from '@/domain/application/repositories/users.repository'
+import { ResourceNotFoundException } from '@/shared/application/exceptions/resource-not-found.exception'
 
 type DeleteAccountRequest = {
   userId: string
@@ -16,6 +17,10 @@ export class DeleteAccountUseCase implements UseCase {
 
   async execute (req: DeleteAccountRequest) {
     const { userId } = req
+    const user = await this.usersRepository.findById(userId)
+    if (!user) {
+      throw new ResourceNotFoundException('User')
+    }
     await this.refreshTokensRepository.deleteManyByUserId(userId)
     await this.usersRepository.delete(userId)
   }
